version bump, split output formating
